+++ src/ez.rs
@@ -1,7 +1,7 @@
 //! An easy API for single-function-call expression evaluation.
 
 use crate::error::Error;
-use crate::parser::parse_noclear;
+use crate::parser::Parser;
 use crate::evaler::Evaler;
 use crate::slab::Slab;
 use crate::evalns::EvalNamespace;
@@ -25,13 +25,13 @@ pub fn ez_eval(expr_str:&str, ns:&mut impl EvalNamespace) -> Result<f64,Error> {
     let mut slab = Slab::new();  // A big block of memory, so we don't need to perform many tiny (and slow!) allocations.
 
     // Here is a one-liner that performs the entire parse-and-eval process:
-    // parse(expr_str, &mut slab.ps)?.from(&slab.ps).eval(&mut slab, &mut ns)
+    // Parser::new().parse(expr_str, &mut slab.ps)?.from(&slab.ps).eval(&mut slab, &mut ns)
 
     // Here is the same process, broken into steps:
 
     // First, parse the string:
     // We use the 'parse_noclear' function instead of 'parse' because we know the slab is empty.
-    let expr_i = parse_noclear(expr_str, &mut slab.ps)?;
+    let expr_i = Parser::new().parse_noclear(expr_str, &mut slab.ps)?;
 
     // 'expr_i' is an index into the Slab.  You can extract the Expression object with either of these:
     //     slab.get_expr(expr_i)  ...OR...  expr_i.from(&slab.ps)