+++ tests/eval.rs
@@ -1,4 +1,4 @@
-use fasteval::{Evaler, Error, Slab, Cached, EmptyNamespace, CachedCallbackNamespace, parse};
+use fasteval::{Evaler, Error, Slab, Cached, EmptyNamespace, CachedCallbackNamespace, Parser};
 use fasteval::bool_to_f64;
 
 use std::mem;
@@ -13,11 +13,11 @@ fn eval() {
     ns.insert("z".to_string(), 3.0);
 
     // Sanity check:
-    assert_eq!(parse("3+3-3/3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 5.0);
+    assert_eq!(Parser::new().parse("3+3-3/3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 5.0);
 
-    assert_eq!(parse("x+y+z", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 6.0);
+    assert_eq!(Parser::new().parse("x+y+z", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 6.0);
 
-    assert_eq!(parse("x+y+z+a", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns), Err(Error::Undefined("a".to_string())));
+    assert_eq!(Parser::new().parse("x+y+z+a", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns), Err(Error::Undefined("a".to_string())));
 }
 
 #[test]
@@ -37,7 +37,7 @@ fn aaa_aaa_sizes() {
 fn aaa_aab_single() {
     let mut slab = Slab::new();
     let mut ns = EmptyNamespace;
-    assert_eq!(parse("123.456", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 123.456f64);
+    assert_eq!(Parser::new().parse("123.456", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap(), 123.456f64);
 }
 
 #[test]
@@ -45,163 +45,163 @@ fn aaa_basics() {
     let mut slab = Slab::new();
 
     assert_eq!(
-        parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap().from(&slab.ps).var_names(&slab),
+        Parser::new().parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap().from(&slab.ps).var_names(&slab),
         BTreeSet::new());
 
     let mut ns = EmptyNamespace;
     assert_eq!(
-        parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(66.66));
     assert_eq!(
-        parse("12.34 + 43.21 - 11.11", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("12.34 + 43.21 - 11.11", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(44.44));
     assert_eq!(
-        parse("11.11 * 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("11.11 * 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(33.33));
     assert_eq!(
-        parse("33.33 / 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("33.33 / 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(11.11));
     assert_eq!(
-        parse("33.33 % 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("33.33 % 3", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.3299999999999983));
     assert_eq!(
-        parse("1 and 2", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1 and 2", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.0));
     assert_eq!(
-        parse("1 && 2", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1 && 2", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.0));
     assert_eq!(
-        parse("2 or 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("2 or 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.0));
     assert_eq!(
-        parse("2 || 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("2 || 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.0));
     assert_eq!(
-        parse("1 > 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1 > 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.0));
     assert_eq!(
-        parse("1 < 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1 < 0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.0));
 
     assert_eq!(
-        parse("+5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("+5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(5.5));
     assert_eq!(
-        parse("-5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("-5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(-5.5));
     assert_eq!(
-        parse("!5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("!5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.0));
     assert_eq!(
-        parse("!0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("!0", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.0));
     assert_eq!(
-        parse("(3 * 3 + 3 / 3)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("(3 * 3 + 3 / 3)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(10.0));
     assert_eq!(
-        parse("(3 * (3 + 3) / 3)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("(3 * (3 + 3) / 3)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(6.0));
 
     assert_eq!(
-        parse("4.4 + -5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("4.4 + -5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(-1.0999999999999996));
     assert_eq!(
-        parse("4.4 + +5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("4.4 + +5.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(9.9));
 
     assert_eq!(
-        parse("x + 1", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("x + 1", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Err(Error::Undefined("x".to_string())));
 
     let mut ns = CachedCallbackNamespace::new(|_,_| Some(3.0));
     assert_eq!(
-        parse("x + 1", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("x + 1", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.0));
 
     assert_eq!(
-        parse("1.2 + int(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + int(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.2));
     assert_eq!(
-        parse("1.2 + ceil(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + ceil(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(5.2));
     assert_eq!(
-        parse("1.2 + floor(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + floor(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.2));
     assert_eq!(
-        parse("1.2 + abs(-3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + abs(-3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.6));
     assert_eq!(
-        parse("1.2 + log(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + log(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.2));
     assert_eq!(
-        parse("1.2 + log(10)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + log(10)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.2));
     assert_eq!(
-        parse("1.2 + log(0)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + log(0)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(std::f64::NEG_INFINITY));
-    assert!(parse("1.2 + log(-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap().is_nan());
+    assert!(Parser::new().parse("1.2 + log(-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns).unwrap().is_nan());
     assert_eq!(
-        parse("1.2 + round(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + round(3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.2));
     assert_eq!(
-        parse("1.2 + round(0.5, 3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + round(0.5, 3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.7));
     assert_eq!(
-        parse("1.2 + round(-3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + round(-3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(-1.8));
     assert_eq!(
-        parse("1.2 + round(0.5, -3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + round(0.5, -3.4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(-2.3));
     assert_eq!(
-        parse("1.2 + min(1,2,0,3.3,-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + min(1,2,0,3.3,-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.19999999999999996));
     assert_eq!(
-        parse("1.2 + min(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + min(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.2));
     assert_eq!(
-        parse("1.2 + max(1,2,0,3.3,-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + max(1,2,0,3.3,-1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(4.5));
     assert_eq!(
-        parse("1.2 + max(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("1.2 + max(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.2));
 
     assert_eq!(
-        parse(r#"12.34 + print ( 43.21, "yay" ) + 11.11"#, &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse(r#"12.34 + print ( 43.21, "yay" ) + 11.11"#, &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(66.66));
 
     assert_eq!(
-        parse("e()", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("e()", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.718281828459045));
     assert_eq!(
-        parse("pi()", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("pi()", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(3.141592653589793));
 
     assert_eq!(
-        parse("sin(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("sin(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.0));
     assert_eq!(
-        parse("cos(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("cos(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.00000000000000006123233995736766));
     assert_eq!(
-        parse("tan(pi()/4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("tan(pi()/4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.9999999999999999));
     assert_eq!(
-        parse("asin(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("asin(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.5707963267948966));
     assert_eq!(
-        parse("acos(0)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("acos(0)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(1.5707963267948966));
     assert_eq!(
-        parse("atan(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("atan(1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.7853981633974483));
     assert_eq!(
-        parse("sinh(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("sinh(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.3012989023072947));
     assert_eq!(
-        parse("cosh(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("cosh(pi()/2)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(2.5091784786580567));
     assert_eq!(
-        parse("tanh(pi()/4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("tanh(pi()/4)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(0.6557942026326724));
 }
 
@@ -233,7 +233,7 @@ fn corners() {
     let mut slab = Slab::new();
     let mut ns = EmptyNamespace;
     assert_eq!(
-        format!("{:?}", parse("(-1) ^ 0.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns)),
+        format!("{:?}", Parser::new().parse("(-1) ^ 0.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns)),
         "Ok(NaN)");
 }
 
@@ -286,44 +286,44 @@ fn custom_func() {
         }
     });
     assert_eq!(
-        parse("x + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(2.5));
 
     assert_eq!(
-        parse("x() + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x() + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(2.5));
 
     assert_eq!(
-        parse("x(1,2,3) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x(1,2,3) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(2.5));
 
     eprintln!("I should see TWO x lookups, 1 y, and 1 z:");
     assert_eq!(
-        parse("x(x,y,z) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x(x,y,z) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(2.5));
 
     eprintln!("I should see TWO x lookups:");
     assert_eq!(
-        parse("x(x,x,x) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x(x,x,x) + 1.5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(2.5));
 
     eprintln!("I should see TWO x lookups:");
     assert_eq!(
-        parse("x(1.0) + x(1.1) + x(1.0) + x(1.1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("x(1.0) + x(1.1) + x(1.0) + x(1.1)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(4.0));
 
     eprintln!("---------------------------");
 
     assert_eq!(
-        parse("foo(1.23)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("foo(1.23)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(12.3));
 
     assert_eq!(
-        parse("bar(1.23, 3.21)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
+        Parser::new().parse("bar(1.23, 3.21)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns}),
         Ok(4.4399999999999995));
 
     assert_eq!(
-        format!("{:?}", parse("bar(1.23)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns})),
+        format!("{:?}", Parser::new().parse("bar(1.23)", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, {ns.cache_clear(); &mut ns})),
         "Ok(NaN)");
 }
 
@@ -342,13 +342,13 @@ fn unsafe_var() {
     let mut ns = EmptyNamespace;
 
     assert_eq!(
-        parse("ua + ub + 5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("ua + ub + 5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(10.79));
 
     ua+=1.0;
     ub+=2.0;
     assert_eq!(
-        parse("ua + ub + 5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
+        Parser::new().parse("ua + ub + 5", &mut slab.ps).unwrap().from(&slab.ps).eval(&slab, &mut ns),
         Ok(13.79));
 
     let _ = (ua,ub);  // Silence compiler warnings about variables not being read.