+++ tests/parse.rs
@@ -1,33 +1,33 @@
-use fasteval::{Error, Slab, parse};
+use fasteval::{Error, Slab, Parser};
 
 #[test]
 fn basics() {
     let mut slab = Slab::new();
-    parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34 + 43.21 + 11.11", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12.34), pairs: [ExprPair(EAdd, EConstant(43.21)), ExprPair(EAdd, EConstant(11.11))] } }, vals:{}, instrs:{} }");
 
-    parse("12.34 + abs ( -43 - 0.21 ) + 11.11", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34 + abs ( -43 - 0.21 ) + 11.11", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(-43.0), pairs: [ExprPair(ESub, EConstant(0.21))] }, 1:Expression { first: EConstant(12.34), pairs: [ExprPair(EAdd, EStdFunc(EFuncAbs(ExpressionI(0)))), ExprPair(EAdd, EConstant(11.11))] } }, vals:{}, instrs:{} }");
 
-    parse("12.34 + abs [ -43 - 0.21 ] + 11.11", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34 + abs [ -43 - 0.21 ] + 11.11", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(-43.0), pairs: [ExprPair(ESub, EConstant(0.21))] }, 1:Expression { first: EConstant(12.34), pairs: [ExprPair(EAdd, EStdFunc(EFuncAbs(ExpressionI(0)))), ExprPair(EAdd, EConstant(11.11))] } }, vals:{}, instrs:{} }");
 
-    parse("12.34 + print ( 43.21 ) + 11.11", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34 + print ( 43.21 ) + 11.11", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(43.21), pairs: [] }, 1:Expression { first: EConstant(12.34), pairs: [ExprPair(EAdd, EPrintFunc(PrintFunc([EExpr(ExpressionI(0))]))), ExprPair(EAdd, EConstant(11.11))] } }, vals:{}, instrs:{} }");
 
-    parse("12.34 + print [ 43.21 ] + 11.11", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34 + print [ 43.21 ] + 11.11", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(43.21), pairs: [] }, 1:Expression { first: EConstant(12.34), pairs: [ExprPair(EAdd, EPrintFunc(PrintFunc([EExpr(ExpressionI(0))]))), ExprPair(EAdd, EConstant(11.11))] } }, vals:{}, instrs:{} }");
 
-    parse("(-1) ^ 0.5", &mut slab.ps).unwrap();
+    Parser::new().parse("(-1) ^ 0.5", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(-1.0), pairs: [] }, 1:Expression { first: EUnaryOp(EParentheses(ExpressionI(0))), pairs: [ExprPair(EExp, EConstant(0.5))] } }, vals:{}, instrs:{} }");
 
-    parse("(1 + [2 - (3 * 4) / 5] ^ 6) % 7", &mut slab.ps).unwrap();
+    Parser::new().parse("(1 + [2 - (3 * 4) / 5] ^ 6) % 7", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(3.0), pairs: [ExprPair(EMul, EConstant(4.0))] }, 1:Expression { first: EConstant(2.0), pairs: [ExprPair(ESub, EUnaryOp(EParentheses(ExpressionI(0)))), ExprPair(EDiv, EConstant(5.0))] }, 2:Expression { first: EConstant(1.0), pairs: [ExprPair(EAdd, EUnaryOp(EParentheses(ExpressionI(1)))), ExprPair(EExp, EConstant(6.0))] }, 3:Expression { first: EUnaryOp(EParentheses(ExpressionI(2))), pairs: [ExprPair(EMod, EConstant(7.0))] } }, vals:{}, instrs:{} }");
 
@@ -37,127 +37,127 @@ fn basics() {
 fn consts() {
     let mut slab = Slab::new();
 
-    parse("12.34", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12.34), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse(".34", &mut slab.ps).unwrap();
+    Parser::new().parse(".34", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.34), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.", &mut slab.ps).unwrap();
+    Parser::new().parse("12.", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    assert_eq!(parse(".", &mut slab.ps), Err(Error::ParseF64(".".to_string())));
+    assert_eq!(Parser::new().parse(".", &mut slab.ps), Err(Error::ParseF64(".".to_string())));
 
-    assert_eq!(parse("12..34", &mut slab.ps), Err(Error::ParseF64("12..34".to_string())));
+    assert_eq!(Parser::new().parse("12..34", &mut slab.ps), Err(Error::ParseF64("12..34".to_string())));
 
-    parse("12.34k", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34k", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12340.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34K", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34K", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12340.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34M", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34M", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12340000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34G", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34G", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12340000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34T", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34T", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(12340000000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34m", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34m", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.01234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34u", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34u", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.00001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34µ", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34µ", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.00001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34n", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34n", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.00000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34p", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34p", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.00000000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34e56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34e56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(1234000000000000000000000000000000000000000000000000000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34e+56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34e+56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(1234000000000000000000000000000000000000000000000000000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34E56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34E56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(1234000000000000000000000000000000000000000000000000000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34E+56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34E+56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(1234000000000000000000000000000000000000000000000000000000.0), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34e-56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34e-56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.0000000000000000000000000000000000000000000000000000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("12.34E-56", &mut slab.ps).unwrap();
+    Parser::new().parse("12.34E-56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.0000000000000000000000000000000000000000000000000000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("+12.34E-56", &mut slab.ps).unwrap();
+    Parser::new().parse("+12.34E-56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(0.0000000000000000000000000000000000000000000000000000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("-12.34E-56", &mut slab.ps).unwrap();
+    Parser::new().parse("-12.34E-56", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(-0.0000000000000000000000000000000000000000000000000000001234), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("-x", &mut slab.ps).unwrap();
+    Parser::new().parse("-x", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EUnaryOp(ENeg(ValueI(0))), pairs: [] } }, vals:{ 0:EStdFunc(EVar(\"x\")) }, instrs:{} }");
 
-    parse("NaN", &mut slab.ps).unwrap();
+    Parser::new().parse("NaN", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(NaN), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("+NaN", &mut slab.ps).unwrap();
+    Parser::new().parse("+NaN", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(NaN), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("-NaN", &mut slab.ps).unwrap();
+    Parser::new().parse("-NaN", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(NaN), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("inf", &mut slab.ps).unwrap();
+    Parser::new().parse("inf", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(inf), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("+inf", &mut slab.ps).unwrap();
+    Parser::new().parse("+inf", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(inf), pairs: [] } }, vals:{}, instrs:{} }");
 
-    parse("-inf", &mut slab.ps).unwrap();
+    Parser::new().parse("-inf", &mut slab.ps).unwrap();
     assert_eq!(format!("{:?}",&slab),
 "Slab{ exprs:{ 0:Expression { first: EConstant(-inf), pairs: [] } }, vals:{}, instrs:{} }");
 
 
 
-    assert_eq!(parse("-infK", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
-    assert_eq!(parse("NaNK", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
-    assert_eq!(parse("12.34e56K", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
+    assert_eq!(Parser::new().parse("-infK", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
+    assert_eq!(Parser::new().parse("NaNK", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
+    assert_eq!(Parser::new().parse("12.34e56K", &mut slab.ps), Err(Error::UnparsedTokensRemaining("K".to_string())));
 
 }
 
@@ -201,7 +201,7 @@ fn unsafe_var() {
         slab.ps.add_unsafe_var("ub".to_string(), &ub);
     }
 
-    parse("ua + ub + 5", &mut slab.ps).unwrap();
+    Parser::new().parse("ua + ub + 5", &mut slab.ps).unwrap();
     assert_eq!(replace_addrs(format!("{:?}",&slab)),
 "Slab{ exprs:{ 0:Expression { first: EStdFunc(EUnsafeVar { name: \"ua\", ptr: 0x???????????? }), pairs: [ExprPair(EAdd, EStdFunc(EUnsafeVar { name: \"ub\", ptr: 0x???????????? })), ExprPair(EAdd, EConstant(5.0))] } }, vals:{}, instrs:{} }");
 }